SNOW-579719 Move static method in ServerConnection to util files
Description
- Move result_set_to_rows and result_set_to_iter to util.py
- Move convert_result_meta_to_attribute to schema_utils.py

Testing
Existing tests
